Verbatim move, no fixes
